+++ TcpStack/TcpSessionMap.cpp
@@ -109,7 +109,7 @@ int CTcpSessionMap::GetCount( )
 	int iCount;
 
 	m_clsMutex.acquire();
-	iCount = m_clsMap.size();
+	iCount = (int)m_clsMap.size();
 	m_clsMutex.release();
 
 	return iCount;
