+++ TestHttp2Client/TestHttp2ClientLoop.cpp
@@ -49,7 +49,7 @@ int TestHttp2ClientLoop( int argc, char * argv[] )
 		memset( szCommand, 0, sizeof(szCommand) );
 		while( fgets( szCommand, sizeof(szCommand), stdin ) )
 		{
-			iLen = strlen( szCommand );
+			iLen = (int)strlen( szCommand );
 			if( iLen >= 2 && szCommand[iLen-2] == '\r' )
 			{
 				szCommand[iLen - 2] = '\0';
